Solve walking into water bug
It happens to be so that ants could try walking into water if certain
conditions were met:

1. The target position of a search is on water tile.
2. The target position has an adjacent tile not being water.

These two conditions together with the fact that
RepeatedAStar.single_search() searches backwards from the target
position to the start position, makes the AvoidWaterFilter miss the
illegal transition, since it only checked next position in an order, and
not the current positon of an order. If the current position happened to
be on a water tile, orders in all directions can be legal, even though
the reversed order onto the water tile was illegal.